Fix CLI path parsing and PSBT helpers
